requirements.txt
fastapi
uvicorn
rapidfuzz
sentence-transformers
llama-cpp-python
